Major change to the way singles are handled.
This is quite a large commit, with major changes to the way Singles are handled, in the same way as done in github.com/gfrd/gfrd. I Should have split it up into multiple commits, since it is a bit hard to read now.

Please ask me if something is unclear. Also see coordinate.py for more comments. I will add other Single (1D/2D) classes later, but first moving onto pairs.

Refactoring. Maybe 'a' should not be a member of Coordinate anymore, but an argument to drawTime and other gf methods. That way no information has to be stored in Coordinate.

Tested against all samples, everything runs, results should not be affected.