Added 1D tests and fixed a typo in Makefile.am
... which lead to errors in compilation at the level
of greens_functions.cpp.

The tests are adapted version from the old code version
but work in the same way as the other tests. They generate
some output which can be used to compare the performance
of the GF with drift vs. the GF without drift and taken out
(or commented out) at a later stage of development.