Simplify logic path to invocation of Green's functions.
Handle escapes nicely in draw_new_com, draw_new_iv and draw_new_position, instead of in greens_function_wrapper.py. I tried to handle this in the Green's functions themselves first, but that was not ideal because an extra random number has to be drawn.

draw_displacement_wrapper is now called draw_r_wrapper, that name was just confusing.

All Green's functions should return position relative to r=r0 on drawR, not relative to r=0. This is only relevant for 1D functions.