Fix bug: update_particle in Multi.hpp did not update
Using map::insert is not what we want when particle has moved.

http://www.cplusplus.com/reference/stl/map/insert/
"Because map containers do not allow for duplicate key values, the insertion
operation checks for each element inserted whether another element exists
already in the container with the same key value, if so, the element is not
inserted and its mapped value is not changed in any way."

Bug introduced November 16th 2010.