change CMake script a little
Even if Boost system or chrono library were not found, it can be used
with MACRO like -DBOOST_CHRONO_HEADER_ONLY. But if user installed boost
as a binary distribution, the source code that would be inlined does not
exist because binary distribution includes source code part as a binary.
So if the libraries are not found, it should be noticed that cmake will
try to compile sgfrd with MACROs that enables to use them as a header
only library because it often generates strange compilation error like
`included files are not found`.