:boom: disable passing polygon to world
Since the constructors of Polygon need to detect which triangles are
connected each other, it uses the information about the periodic
boundary conditions. The condition MUST be the same as that of World
because inconsistency in that part easily makes states of particle
undefined (e.g. inconsistency in position and location on the polygon).
To avoid this, currently simply disable passing a polygon to a world.
You need to initialize world with polygon file.