bug fixes and stability improvements in sGFRD (#432)
* :white_check_mark: replace assertions by exceptions

since sometimes assert()'s are ignored by NDEBUG, like CMake Release
mode. Most of the assertions should be checked in the release mode
because those check depends on the input and other conditions. Replacing
assertions by exceptions, it will never be ignored.

* :recycle: refactor ShellContainer

by using C++11 features

* :sparkles: add ctor that takes Polygon to SGFRDWorld

it is clearly needed.

* :zap: use constexpr specifier for constants

* :recycle: use range-based for to reduce codes

* :art: add missing braces around blocks

* :recycle: use range-for to reduce code size

* :loud_sound: throw exception instead of assertion

since in some cases assert() will be ignored by _NDEBUG

* :art: add ascii-art to show the test condition

* :sparkles: enable to pass triangles to SGFRDFactory

* :children_crossing: bind SGFRDFactory::polygon(triangles)

to python API

* :bug: fix triangle lookup to initialize polygon

to support non-planner polygon correctly

* :white_check_mark: add test codes for polygon::distance

* :recycle: replace branches by boost:algo::clamp

* :recycle: use emplace_back instead of push_back+make_pair

* :white_check_mark: add test case of gaussian-shaped plane

* :bug: fix an edge case on Polygon::distance_sq

For detail, see the ascii arts in the source code.
The previous code does not consider the case 4 and it causes invalid
distance calculation while SGFRD simulation.

There are several stuffs to be done, e.g.
- check case 5.
- refactor Polygon::face_data
  - some of the stuffs become needless because of the changes made in
    this commit.

* :recycle: use unfolded vector, vtop2

Currently it ignores the rotational direction because it does not affect
on the distance calculation. After this, I will add codes to workaround
case 5. While adding the codes, the rotational direction should be
considered because we need to check all the edges in the different
rotational directions.

* :checkered_flag: stop using alternative keywords for MSVC

MSVC does not recognize alternative keywords by default (unless user
includes <ciso646>). Actually, in C, iso646.h defines these keywords
as macros. But in C++, these keywords are built-in functionalities,
so <ciso646> does nothing. MSVC C++ compiler shares most of the part
with C compiler, so this standard incompliance appeared.

* :recycle: modernize SGFRDSimulator::get_intrusive_domains

* :bug: check the shortest path is available

There is an edge case where the shortest path sticks out of the faces.
This change checks the edge lengths and the shortest pass does not cross
it.

* :bug: tighten restrictions on particle and shell size

* :recycle: simplify implementations of get_max_xxx

* :children_crossing: check size limit of particle

* :loud_sound: output an error if particle is too large

* :bug: add missing member variable accessor

* :recycle: redirect F-V/V-V distance functions to F-F

* :zap: remove needless calculation

* :bug: check if face shares vertex

* :rotating_light: be aware of direction even in distance func

* :bug: fix Polygon::direction() function

make it consistent with distance_sq(fp, fp)

* :bug: always check all the pathes

* :boom: throw if particle goes out of the boundary

and if it exceeding the relative tolerance. If the differences are in
the tolerance, then pull back the particles.

* :bug: correct periodic stuff while averaging vtxs

* :bug: consider PBC in geometric restriction

* :mute: remove debug output from SGFRDFactory

* :bug: fix vertex index in geometric restriction

* :loud_sound: diagnosis before finalizing simulator state

* :bug: consider face on which ipv is calculated

since ipv depends on the position of CoM, the direction should be
calculated on the face where CoM locates

* :recycle: add const and refactoring bit

* :mute: remove a needless parameter for logger

* :fire: remove Species in Birth

It is the same as `birth.rule().products().front()`.

* :rotating_light: remove unused variable from SGFRDWorld

* :sparkles: add const/non-const overload to rng access

in SGFRDWorld

* :construction: disable ShellContainer for checking 3D distance

* :alembic: support Birth event experimentally

* :ambulance: add overload for birth to ShellVisitorApplier

* :sparkles: add triangles() API to Polygon

to get a list of triangles

* :bug: clear dt() of SGFRDSimulator

to correctly observe particles

* :bug: check if bursted domain is birth domain

* :bug: apply relative tolerance to check minimum path

the edge inbetween two positions should be long enough to make path
inside the triangles when unfolded

* :recycle: use `relative_tolerance` instead of immediate

* :art: add suffix to include guard

* :art: shorten codes using c++11 features