:fire: remove call_with_size_if_randomly_accessible
In the ancient era, the complexity of the size() method of std::list was
not constant, it was O(N). The purpose of `call_with_size_if_...` seems
that to avoid calling list.size() for the runtime efficiency. It checks
whether the container is randomly accessible and call the passed
function only if the container is randomly accessible (i.e. not an
instance of std::list). After C++11, the complexity of std::list::size
is constant. We no longer need to do this meta programing magic.